fix some dynamic import behavior
